--- a/datalad/support/path.py
+++ b/datalad/support/path.py
@@ -15,6 +15,9 @@ One of the reasons is also to robustify operation with unicode filenames
 import os
 import os.path as op
 
+# to not pollute API importing as _
+from collections import defaultdict as _defaultdict
+
 from functools import wraps
 from itertools import dropwhile
 
@@ -111,3 +114,86 @@ def split_ext(filename):
     file_parts = parts[:1] + tail[::-1]
     ext_parts = parts[1+len(tail):]
     return ".".join(file_parts), "." + ".".join(ext_parts)
+
+
+def get_parent_paths(paths, parents, only_with_parents=False):
+    """Given a list of children paths, return their parent paths among parents
+    or their own path if there is no known parent. A path is also considered its
+    own parent (haven't you watched Predestination?) ;)
+
+    All paths should be POSIX, relative, and not pointing outside (not starting
+    with ../)
+
+    Accent is made on performance to avoid O(len(paths) * len(parents))
+    runtime.  ATM should be typically less than O(len(paths) * len(log(parents)))
+
+    Initial intended use - for a list of paths in the repository
+    to provide their paths as files/submodules known to that repository, to
+    overcome difference in ls-tree and ls-files, where ls-files outputs nothing
+    for paths within submodules.
+    It is coded, so it could later be applied even whenever there are nested
+    parents, e.g. parents = ['sub', 'sub/sub'] and then the "deepest" parent
+    is selected
+
+    Parameters
+    ----------
+    parents: list of str
+    paths: list of str
+    only_with_parents: bool, optional
+      If set to True, return a list of only parent paths where that path had
+      a parent
+
+    Returns
+    -------
+    A sorted list of paths where some entries replaced with their "parents"
+    without duplicates.  So for 'a/b' and 'a/c' with a being among parents, there
+    will be a single 'a'
+    """
+    # Let's do an early check even though then we would skip the checks on paths
+    # being relative etc
+    if not parents:
+        return [] if only_with_parents else paths
+
+    # We will create a lookup for known parent lengths
+    parents = set(parents)  # O(log(len(parents))) lookup
+
+    # rely on path[:n] be quick, and len(parent_lengths) << len(parents)
+    # when len(parents) is large.  We will also bail checking any parent of
+    # the length if at that length path has no directory boundary ('/').
+    #
+    # Create mapping for each length of
+    # parent path to list of parents with that length
+    parent_lengths = _defaultdict(set)
+    for parent in parents:
+        _get_parent_paths_check(parent)
+        parent_lengths[len(parent)].add(parent)
+
+    # Make it ordered in the descending order so we select the deepest/longest parent
+    # and store them as sets for faster lookup.
+    # Could be an ordered dict but no need
+    parent_lengths = [(l, parent_lengths[l]) for l in sorted(parent_lengths, reverse=True)]
+
+    res = set()
+
+    for path in paths:  # O(len(paths)) - unavoidable but could be parallelized!
+        # Sanity check -- should not be too expensive
+        _get_parent_paths_check(path)
+        for parent_length, parents_ in parent_lengths:  # O(len(parent_lengths))
+            if (len(path) < parent_length) or (len(path) > parent_length and path[parent_length] != '/'):
+                continue  # no directory deep enough
+            candidate_parent = path[:parent_length]
+            if candidate_parent in parents_:  # O(log(len(parents))) but expected one less due to per length handling
+                res.add(candidate_parent)
+                break  # it is!
+        else:  # no hits
+            if not only_with_parents:
+                res.add(path)
+
+    return sorted(res)  # TODO: keep it as set?  should we retain original order?
+
+
+def _get_parent_paths_check(path):
+    """A little helper for get_parent_paths"""
+    if isabs(path) or path.startswith(pardir + sep) or path.startswith(curdir + sep):
+        raise ValueError("Expected relative within directory paths, got %r" % path)
+