Intends to at least partially address https://github.com/datalad/datalad/issues/3506
where in a repository with lots of already tracked files, adding more files by providing
their paths would lead only to the heavy CPU load due to paths matching then performed
on DataLad level instead of restricting initial query to  git ls-files  only to the paths
of interest.

Locally I have ran all datalad/core tests and no failures were detected.  This is a PR to see if it would potentially cause some breakage elsewhere.

TODOs
- [ ] make sure nothing is broken
- [x] benchmark