@kyleam 2019-07-02
>  make sure nothing is broken

While you're waiting for Travis to start, here's a failure it will report:

```
 python -m nose -vs datalad/local/tests/test_subdataset.py:test_get_subdatasets
datalad.local.tests.test_subdataset.test_get_subdatasets ... FAIL
[...]

======================================================================
FAIL: datalad.local.tests.test_subdataset.test_get_subdatasets
----------------------------------------------------------------------
Traceback (most recent call last):
  File "/home/kyle/src/python/venvs/datalad/lib/python3.5/site-packages/nose/case.py", line 198, in runTest
    self.test(*self.arg)
  File "/home/kyle/src/python/datalad/datalad/tests/utils.py", line 783, in newfunc
    t(*(arg + (uri,)), **kw)
  File "/home/kyle/src/python/datalad/datalad/local/tests/test_subdataset.py", line 79, in test_get_subdatasets
    'sub dataset1/sub sub dataset1/subm 1',
AssertionError: [] != ['sub dataset1/2', 'sub dataset1/sub sub dataset1', 'sub dataset1/sub sub dataset1/2', 'sub dataset1/sub sub dataset1/subm 1']
-------------------- >> begin captured logging << --------------------
datalad.utils: Level 5: Importing datalad.utils
datalad.utils: DEBUG: Maximal length of cmdline string (adjusted for safety margin): 1252864
datalad.utils: Level 5: Done importing datalad.utils
datalad.cmd: Level 9: Will use git under '/usr/lib/git-annex.linux' (no adjustments to PATH if empty string)
datalad.cmd: Level 9: Running: ['git', 'version']
datalad.cmd: Level 8: Finished running ['git', 'version'] with status 0
datalad.cmd: Level 9: Running: ['git', 'config', '-z', '-l', '--show-origin']
datalad.cmd: Level 8: Finished running ['git', 'config', '-z', '-l', '--show-origin'] with status 0
datalad.ui: Level 5: Starting importing ui
datalad.ui.dialog: Level 5: Starting importing ui.dialog
datalad.ui.dialog: Level 5: Done importing ui.dialog
datalad.ui: Level 5: Initiating UI switcher
datalad.ui: DEBUG: UI set to DialogUI(out=<TextIOWrapper>)
datalad.ui: Level 5: Done importing ui
--------------------- >> end captured logging << ---------------------

----------------------------------------------------------------------
Ran 1 test in 12.685s

FAILED (failures=1)
```