@yarikoptic 2019-07-03
and here is seems only 1 more https://travis-ci.org/datalad/datalad/jobs/553440611

```
FAIL: datalad.support.tests.test_fileinfo.test_subds_path
----------------------------------------------------------------------
Traceback (most recent call last):
  File "/home/travis/virtualenv/python3.5.6/lib/python3.5/site-packages/nose/case.py", line 198, in runTest
    self.test(*self.arg)
  File "/home/travis/virtualenv/python3.5.6/lib/python3.5/site-packages/datalad/tests/utils.py", line 607, in newfunc
    return t(*(arg + (filename,)), **kw)
  File "/home/travis/virtualenv/python3.5.6/lib/python3.5/site-packages/datalad/support/tests/test_fileinfo.py", line 185, in test_subds_path
    assert_equal(stat[subds.repo.pathobj]['state'], 'clean')
AssertionError: 'deleted' != 'clean'
- deleted
+ clean
```