@yarikoptic 2019-07-03
quick summary for the last one (possibly also relates to the previous one on submodules) -- `ls-files` (used for work tree) does not look into submodules for a path within them, whenever `ls-tree` (used for `HEAD`) does

```shell
(git-annex)hopa:~/.tmp/datalad_temp_test_subds_pathyh6f2zwi[master]
$> git ls-files --stage -d -m --exclude-standard sub/some.txt

$> git ls-tree HEAD -r --full-tree -l sub/some.txt           
160000 commit ca339d7604489ad84ba6a3086c7a3846b1869a72       -	sub
```

and even though `--help` for `ls-files` lists `--recurse-submodules` it is not supported:
```
$> git ls-files --recurse-submodules --stage -d -m --exclude-standard sub/some.txt
fatal: ls-files --recurse-submodules unsupported mode

$> git --version
git version 2.22.0.455.g172b71a6c5
```
so we end up with the file within submodule (now we do not sort into submodules first I guess) being considered removed since "known" to HEAD and not to the worktree