@yarikoptic 2019-07-03
`ls-files` though does list submodule itself if submodule path is provided:
```shell
$> git ls-files --stage -d -m --exclude-standard sub/        
160000 ca339d7604489ad84ba6a3086c7a3846b1869a72 0	sub
```