@kyleam 2019-07-31
I've pushed the following updates to this PR:

  * Rebase to deal with conflict.

  * Since we're rebasing anyway, squash some of the existing clean-up commits to their base commit.

  * Use text_type() rather than str() to convert pathlib objects.  See 2f981340c (BF(py2): pathlib: Consistently use text_type for conversion, 2019-04-12).

  * Other minor touch-ups.

  * **Main change:** Move custom submodule logic from subdatasets.py to gitrepo.py, and rewrite GitRepo.get_submodules() to use it.  This should get around the GitPython failures we're seeing in this PR.

I haven't done a close review of the original commits in this PR, but I'm pushing this now so that we can see where we are with the tests.  (The problematic tests from the previous run pass for me locally.)

<details>
<summary>range-diff</summary>

```
 -:  --------- >  1:  ebadb1c42 CLN: subdatasets: Drop unused imports
 -:  --------- >  2:  1256169f4 MV: gitrepo: Absorb custom submodule parser from subdatasets()
 -:  --------- >  3:  ac2544081 DOC: gitrepo: Add docstring for get_submodules_()
 -:  --------- >  4:  fdcdca657 RF: gitrepo: Rewrite get_submodules() to avoid GitPython
 1:  19a4999bf !  5:  ecae1d395 RF: request status only on provided paths (if given)
    @@ -22,7 +22,7 @@
     -                # homogenize wrt subdataset content paths across
     -                # ls-files and ls-tree
     -                None,
    -+                list(map(str, paths)) if paths else None,
    ++                list(map(text_type, paths)) if paths else None,
                      cmd,
                      log_stderr=True,
                      log_stdout=True,
 2:  5976254d3 =  6:  58d34abe9 NF: get_parent_paths to be able to quickly determine within repo paths if paths within submodules provided
 3:  2874f2d82 !  7:  382c7f081 BM: benchmark suite for get_parent_paths
    @@ -2,10 +2,13 @@
     
         BM: benchmark suite for get_parent_paths
     
    - diff --git a/benchmarks/paths.py b/benchmarks/paths.py
    + diff --git a/benchmarks/support/__init__.py b/benchmarks/support/__init__.py
    + new file mode 100644
    +
    + diff --git a/benchmarks/support/path.py b/benchmarks/support/path.py
      new file mode 100644
      --- /dev/null
    - +++ b/benchmarks/paths.py
    + +++ b/benchmarks/support/path.py
     @@
     +# Import functions to be tested with _ suffix and name the suite after the
     +# original function so we could easily benchmark it e.g. by
    @@ -14,7 +17,7 @@
     +
     +from datalad.support.path import get_parent_paths as get_parent_paths_
     +
    -+from .common import SuprocBenchmarks
    ++from ..common import SuprocBenchmarks
     +
     +class get_parent_paths(SuprocBenchmarks):
     +
    @@ -37,13 +40,13 @@
     +        assert get_parent_paths_(self.posixpaths, [], True) == []
     +
     +    def time_one_submod_toplevel(self):
    -+        assert get_parent_paths_(self.posixpaths, ['submod9'], True) == ['submod9']
    ++        get_parent_paths_(self.posixpaths, ['submod9'], True)
     +
     +    def time_one_submod_subdir(self):
    -+        assert get_parent_paths_(self.posixpaths, ['subdir/submod9'], True) == ['subdir/submod9']
    ++        get_parent_paths_(self.posixpaths, ['subdir/submod9'], True)
     +
     +    def time_allsubmods_toplevel_only(self):
    -+        assert get_parent_paths_(self.posixpaths, self.toplevel_submods, True) == self.toplevel_submods
    ++        get_parent_paths_(self.posixpaths, self.toplevel_submods, True)
     +
     +    def time_allsubmods_toplevel(self):
     +        get_parent_paths_(self.posixpaths, self.toplevel_submods)
 4:  176045126 !  8:  237160471 BF: make use of get_parent_paths to mitigate difference in ls-tree and ls-files behavior on paths within submodules
    @@ -19,7 +19,7 @@
                  # convert unconditionally
                  paths = [ut.PurePosixPath(p) for p in paths]
      
    -+        path_strs = list(map(str, paths)) if paths else None
    ++        path_strs = list(map(text_type, paths)) if paths else None
     +
              # this will not work in direct mode, but everything else should be
              # just fine
    @@ -42,7 +42,7 @@
              lgr.debug('Query repo: %s', cmd)
              try:
                  stdout, stderr = self._git_custom_command(
    --                list(map(str, paths)) if paths else None,
    +-                list(map(text_type, paths)) if paths else None,
     +                path_strs,
                      cmd,
                      log_stderr=True,
 -:  --------- >  9:  ba12ba956 DOC: Use public-inbox for Git mailing list link
 5:  5bd64c58e ! 10:  2ac968e72 OPT: do not sort, maintain a set of prior hits
    @@ -15,9 +15,9 @@
         x     3.36±0.04ms      3.18±0.01ms     0.95  paths.get_parent_paths.time_one_submod_toplevel [hopa/virtualenv-py2.7]
         x      4.19±0.2ms      4.04±0.03ms     0.96  paths.get_parent_paths.time_one_submod_toplevel [hopa/virtualenv-py3.7]
     
    - diff --git a/benchmarks/paths.py b/benchmarks/paths.py
    - --- a/benchmarks/paths.py
    - +++ b/benchmarks/paths.py
    + diff --git a/benchmarks/support/path.py b/benchmarks/support/path.py
    + --- a/benchmarks/support/path.py
    + +++ b/benchmarks/support/path.py
     @@
              # and some hierarchy of submodules
              self.nfiles = 40  # per each construct
 6:  b687931a0 <  -:  --------- RF: moved benchmarks/paths.py into benchmarks/support/path.py to reflect main code hierarchy
 7:  b8af9f5ce <  -:  --------- RF: strip assertions from benchmark - they are not tests, painful to maintain while RFing behavior
```

</details>